Move the settings to org.gnome.telepathy-account-widgets
This also deletes all the Empathy-specific settings we don't care about
any more.

Note that moving where we store the settings will lose the settings when
migrating from an old Empathy to the new one using tp-account-widgets.
This is not a real problem as the only setting is the last directory
where an avatar was chosen from (so not a big loss at all).

https://bugzilla.gnome.org/show_bug.cgi?id=699492