tpaw-account-widget: improve the size of the various fields
Fields in the main area now are at least 280px wide. This solves the
problem with fields being too small.

All the fields should expand to fix the available horizontal space. This
solves the problem with fields leaving empty space on their side (in
particular when the advanced section is expanded).