protocol: skip Haze's IRC implementation
Users should be using Idle for IRC, and not Haze. This is implicitly
codified in the widgets array kept in tpaw-account-widget.c. Otherwise,
if Idle is absent we would still be advertising IRC support in the UI,
only to crash immediately afterwards.

All said and done, it is saner to support one single implementation of
a protocol.

https://bugzilla.gnome.org/show_bug.cgi?id=711226