account-widget: Make it work with dialogs using header bars
The tpaw_account_widget_new_for_protocol API has been changed to take
a pointer to a GtkDialog instead of a GtkBox. If one is passed, then
buttons are added to it without directly touching the action area. This
way things will work irrespective of whether a header bar is being used
or not. If NULL is passed, a button box is created and used for packing
the buttons inside the widget as was the case before.

In other words, this only affects users of this API who used to pass
an action area. Those who did not should not be affected.

https://bugzilla.gnome.org/show_bug.cgi?id=730578