Fix GetProtocolsData leak
A GetProtocolsData instance gets associated with a GSimpleAsyncResult in
tpaw_protocol_get_all_async() using
g_simple_async_result_set_op_res_gpointer(). A GDestroyNotify is set so
that the GetProtocolsData instance is freed when the GSimpleAsyncResult
is freed.
However, this is not working as expected as the GetProtocolsData
instance is keeping a reference on the GSimpleAsyncResult it's
associated with, which creates a circular reference and causes the
GetProtocolsData instance to never be freed.

This commit drops the reference on the GSimpleAsyncResult instance owned
by the GetProtocolsData instance as soon as it's no longer needed so
that the GSimpleAsyncResult can be disposed, which triggers the
destruction of the GetProtocolsData too.

https://bugzilla.gnome.org/show_bug.cgi?id=752938